Split test for results and lists
