+++ flutter/realm_flutter/tests/test/realm_test.dart
@@ -8,6 +8,8 @@ import 'package:test_api/src/backend/state.dart' as test_api;
 
 import '../../test/config_test.dart' as config_tests;
 import '../../test/realm_test.dart' as realm_tests;
+import '../../test/list_test.dart' as list_tests;
+import '../../test/results_test.dart' as results_tests;
 
 Future<String> main(List<String> args) async {
   final Completer<String> completer = Completer<String>();
@@ -15,6 +17,8 @@ Future<String> main(List<String> args) async {
 
   await config_tests.main(args);
   await realm_tests.main(args);
+  await list_tests.main(args);
+  await results_tests.main(args);
 
   tearDown(() {
     if (Invoker.current?.liveTest.state.result == test_api.Result.error || Invoker.current?.liveTest.state.result == test_api.Result.failure) {
@@ -25,8 +29,7 @@ Future<String> main(List<String> args) async {
   tearDownAll(() {
     if (failedTests.isNotEmpty) {
       completer.complete(failedTests.join('\n'));
-    }
-    else {
+    } else {
       completer.complete('');
     }
   });
