+++ test/realm_test.dart
@@ -32,89 +32,6 @@ Future<void> main([List<String>? args]) async {
 
   setupTests(Configuration.filesPath, (path) => {Configuration.defaultPath = path});
 
-  group('Configuration tests:', () {
-    test('Configuration can be created', () {
-      Configuration([Car.schema]);
-    });
-
-    test('Configuration exception if no schema', () {
-      expect(() => Configuration([]), throws<RealmException>());
-    });
-
-    test('Configuration default path', () {
-      if (Platform.isAndroid || Platform.isIOS) {
-        expect(Configuration.defaultPath, endsWith(".realm"));
-        expect(Configuration.defaultPath, startsWith("/"), reason: "on Android and iOS the default path should contain the path to the user data directory");
-      } else {
-        expect(Configuration.defaultPath, endsWith(".realm"));
-      }
-    });
-
-    test('Configuration files path', () {
-      if (Platform.isAndroid || Platform.isIOS) {
-        expect(Configuration.filesPath, isNot(endsWith(".realm")), reason: "on Android and iOS the files path should be a directory");
-        expect(Configuration.filesPath, startsWith("/"), reason: "on Android and iOS the files path should be a directory");
-      } else {
-        expect(Configuration.filesPath, equals(""), reason: "on Dart standalone the files path should be an empty string");
-      }
-    });
-
-    test('Configuration get/set path', () {
-      Configuration config = Configuration([Car.schema]);
-      expect(config.path, endsWith('.realm'));
-
-      const path = "my/path/default.realm";
-      config.path = path;
-      expect(config.path, equals(path));
-    });
-
-    test('Configuration get/set schema version', () {
-      Configuration config = Configuration([Car.schema]);
-      expect(config.schemaVersion, equals(0));
-
-      config.schemaVersion = 3;
-      expect(config.schemaVersion, equals(3));
-    });
-
-    test('Configuration readOnly - opening non existing realm throws', () {
-      Configuration config = Configuration([Car.schema], readOnly: true);
-      expect(() => Realm(config), throws<RealmException>("Message: No such table exists"));
-    });
-
-    test('Configuration readOnly - open existing realm with read-only config', () {
-      Configuration config = Configuration([Car.schema]);
-      var realm = Realm(config);
-      realm.close();
-      
-      // Open an existing realm as readonly.
-      config = Configuration([Car.schema], readOnly: true);
-      realm = Realm(config);
-      realm.close();
-    });
-
-    test('Configuration readOnly - reading is possible', () {
-      Configuration config = Configuration([Car.schema]);
-      var realm = Realm(config);
-      realm.write(() => realm.add(Car("Mustang")));
-      realm.close();
-
-      config.isReadOnly = true;
-      realm = Realm(config);
-      var cars = realm.all<Car>();
-      realm.close();
-    });
-
-    test('Configuration readOnly - writing on read-only Realms throws', () {
-      Configuration config = Configuration([Car.schema]);
-      var realm = Realm(config);
-      realm.close();
-
-      config = Configuration([Car.schema], readOnly: true);
-      realm = Realm(config);
-      expect(() => realm.write(() {}), throws<RealmException>("Can't perform transactions on read-only Realms."));
-      realm.close();
-    });
-  });
 
   group('RealmClass tests:', () {
     test('Realm can be created', () {
@@ -441,328 +358,7 @@ Future<void> main([List<String>? args]) async {
       realm.close();
     });
 
-    test('Results.all() should not return null', () {
-      var config = Configuration([Car.schema]);
-      var realm = Realm(config);
-
-      final cars = realm.all<Car>();
-      expect(cars, isNotNull);
-
-      realm.close();
-    });
-
-    test('Results length after deletedMany', () {
-      var config = Configuration([Car.schema]);
-      var realm = Realm(config);
-
-      var cars = realm.all<Car>();
-      expect(cars.length, 0);
-
-      final carOne = Car("Toyota 1");
-      final carTwo = Car("Toyota 2");
-      final carThree = Car("Renault");
-      realm.write(() => realm.addAll([carOne, carTwo, carThree]));
-
-      expect(cars.length, 3);
-
-      final filteredCars = realm.query<Car>('make BEGINSWITH "Toyot"');
-      expect(filteredCars.length, 2);
-
-      realm.write(() => realm.deleteMany(filteredCars));
-      expect(filteredCars.length, 0);
-
-      expect(cars.length, 1);
-
-      realm.close();
-    });
-
-    test('Results length', () {
-      var config = Configuration([Car.schema]);
-      var realm = Realm(config);
-
-      var cars = realm.all<Car>();
-      expect(cars.length, 0);
-
-      final carOne = Car("Toyota");
-      final carTwo = Car("Toyota 1");
-      realm.write(() => realm.addAll([carOne, carTwo]));
-
-      expect(cars.length, 2);
-
-      final filteredCars = realm.query<Car>('make == "Toyota"');
-      expect(filteredCars.length, 1);
-
-      realm.close();
-    });
-
-    test('Results isEmpty', () {
-      var config = Configuration([Car.schema]);
-      var realm = Realm(config);
-
-      var cars = realm.all<Car>();
-      expect(cars.isEmpty, true);
-
-      final car = Car("Opel");
-      realm.write(() => realm.add(car));
-
-      expect(cars.isEmpty, false);
-
-      realm.write(() => realm.delete(car));
-
-      expect(cars.isEmpty, true);
-
-      realm.close();
-    });
-
-    test('Results from query isEmpty', () {
-      var config = Configuration([Dog.schema, Person.schema]);
-      var realm = Realm(config);
-
-      final dogOne = Dog("Pupu", age: 1);
-      final dogTwo = Dog("Ostin", age: 2);
-
-      realm.write(() => realm.addAll([dogOne, dogTwo]));
-
-      var dogs = realm.query<Dog>('age == 0');
-      expect(dogs.isEmpty, true);
-
-      dogs = realm.query<Dog>('age == 1');
-      expect(dogs.isEmpty, false);
-
-      realm.write(() => realm.deleteMany(dogs));
-      expect(dogs.isEmpty, true);
-
-      dogs = realm.all<Dog>();
-      expect(dogs.isEmpty, false);
-
-      realm.close();
-    });
-
-    test('Results get by index', () {
-      var config = Configuration([Car.schema]);
-      var realm = Realm(config);
-
-      final car = Car('');
-      realm.write(() => realm.add(car));
-
-      final cars = realm.all<Car>();
-      expect(cars[0].make, car.make);
-
-      realm.close();
-    });
-
-    group('Query and sort tests:', () {
-      test('Query results', () {
-        var config = Configuration([Car.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm
-          ..add(Car("Audi"))
-          ..add(Car("Tesla")));
-        final cars = realm.all<Car>().query('make == "Tesla"');
-        expect(cars.length, 1);
-        expect(cars[0].make, "Tesla");
-
-        realm.close();
-      });
-
-      test('Query class', () {
-        var config = Configuration([Car.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm
-          ..add(Car("Audi"))
-          ..add(Car("Tesla")));
-        final cars = realm.query<Car>('make == "Tesla"');
-        expect(cars.length, 1);
-        expect(cars[0].make, "Tesla");
-
-        realm.close();
-      });
-
-      test('Query results with parameter', () {
-        var config = Configuration([Car.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm
-          ..add(Car("Audi"))
-          ..add(Car("Tesla")));
-        final cars = realm.all<Car>().query(r'make == $0', ['Tesla']);
-        expect(cars.length, 1);
-        expect(cars[0].make, "Tesla");
-
-        realm.close();
-      });
-
-      test('Query results with multiple parameters', () {
-        var config = Configuration([Team.schema, Person.schema]);
-        var realm = Realm(config);
-
-        final p1 = Person('p1');
-        final p2 = Person('p2');
-        final t1 = Team("A1", players: [p1]); // match
-        final t2 = Team("A2", players: [p2]); // correct prefix, but wrong player
-        final t3 = Team("B1", players: [p1, p2]); // wrong prefix, but correct player
-
-        realm.write(() => realm.addAll([t1, t2, t3]));
-
-        expect(t1.players, [p1]);
-        expect(t2.players, [p2]);
-        expect(t3.players, [p1, p2]);
-
-        final filteredTeams = realm.all<Team>().query(r'$0 IN players AND name BEGINSWITH $1', [p1, 'A']);
-        expect(filteredTeams.length, 1);
-        expect(filteredTeams, [t1]);
-
-        realm.close();
-      });
-
-      test('Query class with parameter', () {
-        var config = Configuration([Car.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm
-          ..add(Car("Audi"))
-          ..add(Car("Tesla")));
-        final cars = realm.query<Car>(r'make == $0', ['Tesla']);
-        expect(cars.length, 1);
-        expect(cars[0].make, "Tesla");
-
-        realm.close();
-      });
-
-      test('Query class with multiple parameters', () {
-        var config = Configuration([Team.schema, Person.schema]);
-        var realm = Realm(config);
-
-        final p1 = Person('p1');
-        final p2 = Person('p2');
-        final t1 = Team("A1", players: [p1]);
-        final t2 = Team("A2", players: [p2]);
-        final t3 = Team("B1", players: [p1, p2]);
-
-        realm.write(() => realm
-          ..add(t1)
-          ..add(t2)
-          ..add(t3));
-
-        expect(t1.players, [p1]);
-        expect(t2.players, [p2]);
-        expect(t3.players, [p1, p2]);
-        final filteredTeams = realm.query<Team>(r'$0 IN players AND name BEGINSWITH $1', [p1, 'A']);
-        expect(filteredTeams.length, 1);
-        expect(filteredTeams[0].name, "A1");
-
-        realm.close();
-      });
-
-      test('Query results with no arguments throws', () {
-        var config = Configuration([Car.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm.add(Car("Audi")));
-        expect(() => realm.all<Car>().query(r'make == $0'), throws<RealmException>("no arguments are provided"));
-
-        realm.close();
-      });
-
-      test('Query results with wrong argument types (int for string) throws', () {
-        var config = Configuration([Car.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm.add(Car("Audi")));
-        expect(() => realm.all<Car>().query(r'make == $0', [1]), throws<RealmException>("Unsupported comparison between type"));
-        realm.close();
-      });
-
-      test('Query results with wrong argument types (bool for int) throws ', () {
-        var config = Configuration([Dog.schema, Person.schema]);
-        var realm = Realm(config);
-        realm.write(() => realm.add(Dog("Foxi")));
-        expect(() => realm.all<Dog>().query(r'age == $0', [true]), throws<RealmException>("Unsupported comparison between type"));
-        realm.close();
-      });
-
-      test('Query list', () {
-        final config = Configuration([Team.schema, Person.schema]);
-        final realm = Realm(config);
-
-        final person = Person('John');
-        final team = Team('team1', players: [
-          Person('Pavel'),
-          person,
-          Person('Alex'),
-        ]);
-
-        realm.write(() => realm.add(team));
-
-        // TODO: Get rid of cast, once type signature of team.players is a RealmList<Person>
-        // as opposed to the List<Person> we have today.
-        final result = (team.players as RealmList<Person>).query(r'name BEGINSWITH $0', ['J']);
-
-        expect(result, [person]);
-
-        realm.close();
-      });
-
-      test('Sort result', () {
-        var config = Configuration([Person.schema]);
-        var realm = Realm(config);
-
-        realm.write(() => realm.addAll([
-              Person("Michael"),
-              Person("Sebastian"),
-              Person("Kimi"),
-            ]));
-
-        final result = realm.query<Person>('TRUEPREDICATE SORT(name ASC)');
-        final resultNames = result.map((p) => p.name).toList();
-        final sortedNames = [...resultNames]..sort();
-        expect(resultNames, sortedNames);
-        realm.close();
-      });
-
-      test('Sort order preserved under db ops', () {
-        var config = Configuration([Dog.schema, Person.schema]);
-        var realm = Realm(config);
-
-        final dog1 = Dog("Bella", age: 1);
-        final dog2 = Dog("Fido", age: 2);
-        final dog3 = Dog("Oliver", age: 3);
-
-        realm.write(() => realm.addAll([dog1, dog2, dog3]));
-        var result = realm.query<Dog>('TRUEPREDICATE SORT(name ASC)');
-        final snapshot = result.toList();
-
-        expect(result, orderedEquals(snapshot));
-        expect(result.map((d) => d.name), snapshot.map((d) => d.name));
-        result = realm.query<Dog>('TRUEPREDICATE SORT(name ASC)'); // redoing query won't change that
-        expect(result, orderedEquals(snapshot));
-        expect(result.map((d) => d.name), snapshot.map((d) => d.name));
-
-        realm.write(() => realm.delete(dog1)); // result will update, snapshot will not, but an object has died
-
-        expect(() => snapshot[0].name, throws<RealmException>());
-        snapshot.removeAt(0); // remove dead object
-
-        expect(result, orderedEquals(snapshot));
-        expect(result.map((d) => d.name), snapshot.map((d) => d.name));
-
-        realm.write(() => realm.add(Dog("Bella", age: 4)));
-
-        expect(result, isNot(orderedEquals(snapshot)));
-        expect(result, containsAllInOrder(snapshot));
-        realm.close();
-      });
-    });
-    test('Lists create object with a list property', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      final team = Team("Ferrari");
-      realm.write(() => realm.add(team));
-
-      final teams = realm.all<Team>();
-      expect(teams.length, 1);
-      expect(teams[0].name, "Ferrari");
-      expect(teams[0].players, isNotNull);
-      expect(teams[0].players.length, 0);
-      realm.close();
-    });
+   
 
     group('notification', () {
       test('Results notifications', () async {
@@ -940,225 +536,6 @@ Future<void> main([List<String>? args]) async {
       expect(teams[0].scores[0], 1);
       expect(teams[0].scores[1], 2);
       expect(teams[0].scores[2], 3);
-      realm.close();
-    });
-
-    test('Lists get set', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      final team = Team("Ferrari");
-      realm.write(() => realm.add(team));
-
-      final teams = realm.all<Team>();
-      expect(teams.length, 1);
-      final players = teams[0].players;
-      expect(players, isNotNull);
-      expect(players.length, 0);
-
-      realm.write(() => players.add(Person("Michael")));
-      expect(players.length, 1);
-
-      realm.write(() => players.addAll([
-            Person("Sebastian"),
-            Person("Kimi"),
-          ]));
-
-      expect(players.length, 3);
-
-      expect(players[0].name, "Michael");
-      expect(players[1].name, "Sebastian");
-      expect(players[2].name, "Kimi");
-      realm.close();
-    });
-
-    test('Lists get invalid index throws exception', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      final team = Team("Ferrari");
-      realm.write(() => realm.add(team));
-
-      final teams = realm.all<Team>();
-      final players = teams[0].players;
-
-      expect(() => players[-1], throws<RealmException>("Index out of range"));
-      expect(() => players[800], throws<RealmException>());
-      realm.close();
-    });
-
-    test('Lists set invalid index throws', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      final team = Team("Ferrari");
-      realm.write(() => realm.add(team));
-
-      final teams = realm.all<Team>();
-      final players = teams[0].players;
-
-      expect(() => realm.write(() => players[-1] = Person('')), throws<RealmException>("Index out of range"));
-      expect(() => realm.write(() => players[800] = Person('')), throws<RealmException>());
-      realm.close();
-    });
-
-    test('List clear items from list', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      //Create a team
-      final team = Team("Team");
-      realm.write(() => realm.add(team));
-
-      //Add players to the team
-      final newPlayers = [
-        Person("Michael Schumacher"),
-        Person("Sebastian Vettel"),
-        Person("Kimi Räikkönen"),
-      ];
-
-      realm.write(() {
-        team.players.addAll(newPlayers);
-      });
-
-      //Ensure teams and players are in realm
-      var teams = realm.all<Team>();
-      expect(teams.length, 1);
-
-      var players = teams[0].players;
-      expect(players, isNotNull);
-      expect(players.length, 3);
-
-      //Clear list of team players
-      realm.write(() => teams[0].players.clear());
-
-      expect(teams[0].players.length, 0);
-
-      //Ensure that players objects still exist in realm detached from the team
-      final allPlayers = realm.all<Person>();
-      expect(allPlayers.length, 3);
-      realm.close();
-    });
-
-    test('List clear - same list related to two objects', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      //Create two teams
-      final teamOne = Team("TeamOne");
-      final teamTwo = Team("TeamTwo");
-      realm.write(() {
-        realm.add(teamOne);
-        realm.add(teamTwo);
-      });
-
-      //Create common players list for both teams
-      final newPlayers = [
-        Person("Michael Schumacher"),
-        Person("Sebastian Vettel"),
-        Person("Kimi Räikkönen"),
-      ];
-      realm.write(() {
-        teamOne.players.addAll(newPlayers);
-        teamTwo.players.addAll(newPlayers);
-      });
-
-      //Ensure that teams and players exist in realm
-      var teams = realm.all<Team>();
-      expect(teams.length, 2);
-      expect(teams[0].players, isNotNull);
-      expect(teams[0].players.length, 3);
-      expect(teams[1].players, isNotNull);
-      expect(teams[1].players.length, 3);
-
-      //Clear first team's players only
-      realm.write(() => teams[0].players.clear());
-
-      //Ensure that second team is still related to players
-      expect(teams[0].players.length, 0);
-      expect(teams[1].players.length, 3);
-
-      //Ensure players still exist in realm
-      final players = realm.all<Person>();
-      expect(players.length, 3);
-      realm.close();
-    });
-
-    test('List clear - same item added to two lists', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      //Create two Teams
-      final teamOne = Team("TeamOne");
-      final teamTwo = Team("TeamTwo");
-      realm.write(() {
-        realm.add(teamOne);
-        realm.add(teamTwo);
-      });
-
-      //Add the same player to both teams
-      Person player = Person("Michael Schumacher");
-      realm.write(() {
-        teamOne.players.add(player);
-        teamTwo.players.add(player);
-      });
-
-      //Ensure teams and player are in realm
-      var teams = realm.all<Team>();
-      expect(teams.length, 2);
-      expect(teams[0].players, isNotNull);
-      expect(teams[0].players.length, 1);
-      expect(teams[1].players, isNotNull);
-      expect(teams[1].players.length, 1);
-
-      //Clear player from the first team
-      realm.write(() => teams[0].players.clear());
-
-      //Ensure that the second team has no more players
-      // but the first team is still related to the player
-      expect(teams[0].players.length, 0);
-      expect(teams[1].players.length, 1);
-
-      //Ensure the player still exists in realm
-      final allPlayers = realm.all<Person>();
-      expect(allPlayers.length, 1);
-      realm.close();
-    });
-
-    test('List clear in closed realm - expected exception', () {
-      var config = Configuration([Team.schema, Person.schema]);
-      var realm = Realm(config);
-
-      //Create a team
-      var team = Team("TeamOne");
-      realm.write(() => realm.add(team));
-
-      //Add the player to the team
-      realm.write(() => team.players.add(Person("Michael Schumacher")));
-
-      //Ensure teams and player are in realm
-      var teams = realm.all<Team>();
-      expect(teams.length, 1);
-      expect(teams[0].players, isNotNull);
-      expect(teams[0].players.length, 1);
-
-      var players = teams[0].players;
-
-      realm.close();
-      expect(
-          () => realm.write(() {
-                players.clear();
-              }),
-          throws<RealmException>());
-
-      realm = Realm(config);
-
-      //Teams must be reloaded since realm was reopened
-      teams = realm.all<Team>();
-
-      //Ensure that the team is still related to the player
-      expect(teams.length, 1);
-      expect(teams[0].players.length, 1);
       realm.close();
     });
